modules/PersonLookupService/requirements.txt
grpcio==1.39.0
grpcio-tools==1.39.0
requests==2.25.0

